move the new http server to the existing package
